remove support of netstandard2.0
