+++ .github/workflows/nuget.yaml
@@ -13,10 +13,6 @@ jobs:
     - uses: actions/checkout@v1
       with:
         fetch-depth: 0
-    - name: Setup dotnet SDK 2.1
-      uses: actions/setup-dotnet@v1
-      with:
-        dotnet-version: '2.1.x'
     - name: Setup dotnet SDK 3.1
       uses: actions/setup-dotnet@v1
       with:
