+++ tests/KubernetesClient.Tests/WatchTests.cs
@@ -6,6 +6,7 @@ using System.IO;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
+using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using k8s.Models;
@@ -647,6 +648,40 @@ namespace k8s.Tests
                     await client.ListNamespacedPodWithHttpMessagesAsync("default", watch: true,
                         cancellationToken: cts.Token).ConfigureAwait(false);
                 }).ConfigureAwait(false);
+            }
+        }
+
+        [Fact]
+        public void ReadError()
+        {
+            var data = Encoding.UTF8.GetBytes(
+                "{\"type\":\"ERROR\",\"object\":{\"kind\":\"Status\",\"apiVersion\":\"v1\",\"metadata\":{},\"status\":\"Failure\",\"message\":\"too old resource version: 44982(53593)\",\"reason\":\"Gone\",\"code\":410}}");
+
+            using (var stream = new MemoryStream(data))
+            using (var reader = new StreamReader(stream))
+            {
+                Exception recordedException = null;
+                var mre = new ManualResetEvent(false);
+
+                var watcher = new Watcher<V1Pod>(
+                    () => Task.FromResult(reader),
+                    null,
+                    (exception) =>
+                    {
+                        recordedException = exception;
+                        mre.Set();
+                    });
+
+                mre.WaitOne();
+
+                Assert.NotNull(recordedException);
+
+                var k8sException = recordedException as KubernetesException;
+
+                Assert.NotNull(k8sException);
+                Assert.NotNull(k8sException.Status);
+                Assert.Equal("too old resource version: 44982(53593)", k8sException.Message);
+                Assert.Equal("too old resource version: 44982(53593)", k8sException.Status.Message);
             }
         }
     }
